refactor(utils): move individual utils to separate files
