+++ src/utils/arrays/find-last.ts
@@ -1,0 +1,10 @@
+export function findLast<T>(
+  items: T[],
+  predicate: (item: T) => boolean
+): T | undefined {
+  for (let index = items.length - 1; index >= 0; index--) {
+    if (predicate(items[index])) {
+      return items[index];
+    }
+  }
+}
