+++ src/utils/arrays/flatten.test.ts
@@ -1,13 +1,5 @@
-import { wizards } from "../test-data";
-import { findLast, flatten, shallowEqual } from "./array";
-
-test("shallowEqual", () => {
-  expect(shallowEqual(["hello", "world"], ["hello", "world"])).toBe(true);
-  expect(shallowEqual(["hello", "world"], ["hello"])).toBe(false);
-  expect(shallowEqual(["hello", "world"], ["hello", "world", "lorem"])).toBe(
-    false
-  );
-});
+import { wizards } from "../../test-data";
+import { flatten } from "./flatten";
 
 test("flatten", () => {
   expect(
@@ -48,14 +40,4 @@ test("flatten", () => {
     { op: "remove", path: '/{"id":7}/0' },
     { op: "remove", path: '/{"id":8}/0' },
   ]);
-});
-
-test("findLast", () => {
-  expect(
-    findLast(wizards, (wizard) => wizard.house === "gryffindor")
-  ).toMatchObject({ name: "ron" });
-
-  expect(
-    findLast(wizards, (wizard) => wizard.house === "hufflepuff")
-  ).toBeUndefined();
 });
