+++ src/utils/arrays/index.ts
@@ -1,0 +1,4 @@
+export { findLast } from "./find-last";
+export { flatten } from "./flatten";
+export { keyArrayBy } from "./key-array-by";
+export { shallowEqual } from "./shallow-equal";
