+++ src/utils/arrays/shallow-equal.ts
@@ -1,0 +1,6 @@
+export function shallowEqual(left: unknown[], right: unknown[]) {
+  return (
+    left.length === right.length &&
+    left.every((item, index) => item === right[index])
+  );
+}
