+++ src/utils/collections/index.ts
@@ -1,0 +1,5 @@
+export { cloneDeep } from "./clone-deep";
+export { dot } from "./dot";
+export { isEqual } from "./is-equal";
+export { set } from "./set";
+export { unset } from "./unset";
