+++ src/utils/collections/is-equal.ts
@@ -1,0 +1,4 @@
+// @todo placeholder isEqual function
+export function isEqual(left: unknown, right: unknown) {
+  return JSON.stringify(left) === JSON.stringify(right);
+}
