+++ src/utils/collections/set.test.ts
@@ -1,0 +1,5 @@
+import { set } from "./set";
+
+test("set", () => {
+  expect(set({}, "a.b.c", "hello")).toEqual({ a: { b: { c: "hello" } } });
+});
