+++ src/utils/collections/unset.test.ts
@@ -1,0 +1,7 @@
+import { unset } from "./unset";
+
+test("unset", () => {
+  expect(unset({ a: { b: { c: "hello", d: {} } } }, "a.b.c")).toEqual({
+    a: { b: { d: {} } },
+  });
+});
