+++ src/utils/types/as-array.ts
@@ -1,0 +1,10 @@
+import { isArray } from "./is-array";
+import { isNotNullOrUndefined } from "./is-not-null-or-undefined";
+
+export function asArray<T>(value: T | T[] | null | undefined): T[] {
+  if (!isNotNullOrUndefined(value)) {
+    return [];
+  }
+
+  return isArray(value) ? value : [value];
+}
