+++ src/utils/types/is-array.ts
@@ -1,0 +1,3 @@
+export function isArray<T>(value: unknown): value is T[] {
+  return Array.isArray(value);
+}
