+++ src/utils/types/is-defined.test.ts
@@ -1,0 +1,10 @@
+import { isDefined } from "./is-defined";
+
+test.each([
+  [1, true],
+  ["test", true],
+  [undefined, false],
+])("isDefined(%s) is %s", (input, expected) => {
+  const actual = isDefined(input);
+  expect(actual).toBe(expected);
+});
