+++ src/utils/types/is-defined.ts
@@ -1,0 +1,3 @@
+export function isDefined<T>(value: T | undefined): value is T {
+  return value !== undefined;
+}
