+++ src/utils/types/is-multi-dimensional-array.ts
@@ -1,0 +1,6 @@
+import { isArray } from "./is-array";
+import { isPopulatedArray } from "./is-populated-array";
+
+export function isMultiDimensionalArray<T>(value: unknown): value is T[][] {
+  return isPopulatedArray(value) && isArray(value[0]);
+}
