+++ src/utils/types/is-not-null-or-undefined.ts
@@ -1,0 +1,8 @@
+import { isDefined } from "./is-defined";
+import { isNotNull } from "./is-not-null";
+
+export function isNotNullOrUndefined<T>(
+  value: T | null | undefined
+): value is T {
+  return isDefined(value) && isNotNull(value);
+}
