+++ src/utils/types/is-not-null.ts
@@ -1,0 +1,3 @@
+export function isNotNull<T>(value: T | null): value is T {
+  return value !== null;
+}
