+++ src/utils/types/is-null.ts
@@ -1,0 +1,3 @@
+export function isNull<T>(value: T | null): value is null {
+  return value === null;
+}
