+++ src/utils/types/is-number.ts
@@ -1,0 +1,3 @@
+export function isNumber(value: unknown): value is number {
+  return typeof value === "number";
+}
