+++ src/utils/types/is-numeric.ts
@@ -1,0 +1,6 @@
+export function isNumeric(value: unknown) {
+  return (
+    typeof value === "number" ||
+    (typeof value === "string" && parseInt(value, 10).toString() === value)
+  );
+}
