+++ src/utils/types/is-object-of-properties.ts
@@ -1,0 +1,14 @@
+import { shallowEqual } from "../arrays";
+import { isArray } from "./is-array";
+
+export function isObjectOfProperties<T>(
+  object: unknown,
+  properties: string[]
+): object is T {
+  return (
+    object !== null &&
+    typeof object === "object" &&
+    !isArray(object) &&
+    shallowEqual(Object.keys(object), properties)
+  );
+}
