+++ src/utils/types/is-object.test.ts
@@ -1,0 +1,11 @@
+import { isObject } from "./is-object";
+
+test.each([
+  [1, false],
+  ["test", false],
+  [[], false],
+  [{}, true],
+])("isObject(%s) is %s", (input, expected) => {
+  const actual = isObject(input);
+  expect(actual).toBe(expected);
+});
