+++ src/utils/types/is-object.ts
@@ -1,0 +1,3 @@
+export function isObject(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && !Array.isArray(value);
+}
