+++ src/utils/types/is-populated-array.ts
@@ -1,0 +1,5 @@
+import { isArray } from "./is-array";
+
+export function isPopulatedArray<T, Y>(value: T[] | Y): value is T[] {
+  return isArray(value) && value.length > 0;
+}
