+++ src/utils/types/is-scalar.ts
@@ -1,0 +1,3 @@
+export function isScalar(value: unknown): value is string | number | boolean {
+  return ["number", "string", "boolean"].includes(typeof value);
+}
