+++ src/utils/types/is-single-array.ts
@@ -1,0 +1,5 @@
+import { isArray } from "./is-array";
+
+export function isSingleArray<T, Y>(value: T[] | Y): value is [T] {
+  return isArray(value) && value.length === 1;
+}
