+++ src/utils/types/is-singular.ts
@@ -1,0 +1,5 @@
+import { isArray } from "./is-array";
+
+export function isSingular<T>(value: unknown): value is T {
+  return !isArray(value);
+}
