+++ src/utils/types/is-undefined.ts
@@ -1,0 +1,3 @@
+export function isUndefined<T>(value: T | undefined): value is undefined {
+  return value === undefined;
+}
